Fix PRNGs

+++ src/main/java/org/inasayaflanderin/abyssine/miscellaneous/rng/ACG.java
@@ -6,11 +6,11 @@ public class ACG extends LCG {
     @Serial
     private static final long serialVersionUID = -5223182888290600404L;
 
-    public ACG(double increment, double modulus) {
+    public ACG(long increment, long modulus) {
         super(1, increment, modulus);
     }
 
-    public ACG(long seed, double increment, double modulus) {
+    public ACG(long seed, long increment, long modulus) {
         super(seed, 1, increment, modulus);
     }
 }
