+++ src/main/java/org/inasayaflanderin/abyssine/miscellaneous/rng/MCG.java
@@ -6,11 +6,11 @@ public class MCG extends LCG {
     @Serial
     private static final long serialVersionUID = 7837947318768033003L;
 
-    public MCG(double multiplier, double modulus) {
+    public MCG(long multiplier, long modulus) {
         super(multiplier, 0, modulus);
     }
 
-    public MCG(long seed, double multiplier, double modulus) {
+    public MCG(long seed, long multiplier, long modulus) {
         super(seed, multiplier, 0, modulus);
     }
 }
