+++ src/main/java/org/inasayaflanderin/abyssine/miscellaneous/rng/MWC.java
@@ -6,11 +6,11 @@ public class MWC extends LMWC {
     @Serial
     private static final long serialVersionUID = 617819207934434494L;
 
-    public MWC(double multiplier, double increment, double modulus) {
+    public MWC(long multiplier, long increment, long modulus) {
         super(multiplier, increment, modulus, 1);
     }
 
-    public MWC(double seed, double multiplier, double increment, double modulus) {
+    public MWC(long seed, long multiplier, long increment, long modulus) {
         super(seed, multiplier, increment, modulus, 1);
     }
 }
